fix: move prechat-tasks and prechat-task-icons to top-level shared dirs
Moves cross-domain shared constants out of domains/onboarding/ to avoid
no-cross-domain-imports lint violations:
- prechat-tasks.ts → src/types/ (pure data catalog)
- prechat-task-icons.ts → src/components/ (icon mapping)

Both files are consumed by onboarding and chat domains.
